backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.1.1
itsdangerous==2.2.0
cryptography==45.0.5
